refactor(play): extract setup-validation utils (S26.0c)
Troisieme tranche du refactor S26.0 :
play/[id]/page.tsx 1514 -> 1448 lignes (-66, total -218 depuis 1666).

- Extrait getMySide() (3 callsites) et validatePlacement() (2 callsites,
  ~75 lignes) dans apps/web/app/play/[id]/utils/setup-validation.ts.
- Les fonctions sont maintenant pures (plus d'access a la closure
  React) :
  - getMySide(ext, teamNameA, teamNameB) prend les noms d'equipe en
    args (`string | null | undefined` pour matcher l'usage en page).
  - validatePlacement(ext, playerId, pos, mySide) prend mySide
    precalcule en arg.
- showSetupError reste inline (dependance forte a setSetupError React).

Slices restantes pour S26.0 : reste de la phase setup (handleDrop,
onCellClick, handleValidatePlacement, ~280 lignes), kickoff handlers,
BlockChoice/ThrowTeamMate flows, types Move unions pour eliminer
les `as any` (13 occurrences).
